Require that framebuffer attachments are cube complete
This rule is added to improve the portability of WebGL implementations -
some underlying GL implementations enforce this restriction found in
newer OpenGL specs and working around it is complex. The backwards
compatibility impact of enforcing this rule in WebGL is expected to be
minimal.

This rule is not extended to require mipmap completeness as well, as
using non-mipmap complete attachments is already covered by WebGL tests
and working, and there might be content relying on that.