Test availability of queries' results per WebGL 2.0 spec rules.
Queries' results must not become available until control returns to the browser's event loop. Verify this for ANY_SAMPLES_PASSED and ANY_SAMPLES_PASSED_CONSERVATIVE occlusion queries in a new targeted test, and modify the transform_feedback test to verify this for TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN queries (which require more setup code).

Change the ported dEQP transform feedback test to enforce that the query's result is as expected, since the logic was recently changed to obey the new specification rules. Fix a bug in the computation of the expected number of primitives.