Restrict mag filter to NEAREST for float type texture
Float type texture isn't texture filterable in ES 3.0.
A texture isn't complete if its internal color format is not
texture-filterable and either the magnification filter is not NEAREST or
the minification filter is neigher NEAREST nor NEAREST_MIPMAP_NEAREST.
Chrome will use a black texture for texture being accessed is not
complete which causes tex-storage-2d.html to fail.