Add disallowed variants of ESSL3 operators to spec
Supporting ternary and sequence operators on void and arrays is likely to
be made optional in GLSL ES 3.00 according to current draft for a future
version of the spec. For WebGL, it's the best to disallow any features
that may not be implemented on all reasonable target platforms. These
restrictions on target platforms could be worked around, but that adds an
extra burden on implementations for very little gain in corner case
features in this case.