The global invariant pragma is illegal in fragment shaders.
GLSL ES 3.00.4 section 4.6.1 page 53 states:
 "It is an error to use this pragma in a fragment shader."

Changed the tests which stated that the fragment shader should compile
successfully to instead state that it should fail to compile, and
change one test that stated that the link should succeed to say that
it should fail.

This causes Chrome to start failing one of these tests on NVIDIA's
drivers, but that's because the shader translator isn't properly
checking this, which needs to be fixed.