Add ImageBitmap to TexImageSource
This change adds ImageBitmap support to texImage2D,
texSubImage2D (WebGL 1 spec), and texImage3D (WebGL 2 spec).
The origin restrictions logic was updated to take
into account the origin-clean flag of an ImageBitmap's bitmap.
The terminology for the origin-clean flag was tweaked to match the
language used in the HTML spec: "the origin-clean flag of an
HTMLCanvasElement" becomes "the origin-clean flag of an
HTMLCanvasElement's bitmap".