Make some uses of "defined" shader preprocessor errors
The GLES working group decided in recent discussions that these cases
should be errors, but left the spec undefined in older versions of the
API. WebGL should specify them to be errors to be consistent with the
direction GLES is going.

The changes to the tests will be upstreamed to native dEQP as well.