Make timestamps optional in timestamp WebGL test.
Allow the counter bits for timestamps to be 0 for the WebGL
EXT_disjoint_timer_query conformance test and disable the timestamp
portion of the test if that is the case. Certain backing drivers for
WebGL such as the OSX drivers and ANGLE on D3D11 do not support
timestamps and it is within the spec to disable timestamp functionality
in the extension by setting the number of counter bits for the
timestamps to be 0. This updates the test to check if the timestamp
counter bits is set to 0 and to skip the timestamp portion of the test
if that is the case.