Performing some optimization on tcuFloat.
In several places repeated allocations of tcuFloat or tcuFloatDescripton were
reduced to lessen memory churn, especially with tcuFloat, which allocates a
TypedArray (in some uses multiple times), which is a very expensive allocation.

Also replaced tcuFloat.numberToHalfFloat with a version that does the conversion
inline rather than relying on the tcuFloat class. This speeds up some tests
nearly 100x!