Set default context version for WebGL 1.0 and 2.0 texture tests.
This allows these tests to be run as the test harness would do so without the ?webglVersion query string, while still allowing that query string to override the default version. In this way, the WebGL 1.0 tests can still be run against WebGL 2.0 contexts.

This is a follow-on to #1556, which made a similar change to the rest of the WebGL 2.0 conformance tests.