Shorten texture test filenames.
Some of these file names are over 255 characters which is (still, in the year 2016) causing problems on Windows machines.

Remove the element type from the test's file name; it's already present in the directory containing the test. Also remove the image-and-sub-image string. These will comfortably get all of the test names under this limit.