Replace drawQuads with drawTriangles
The dEQP reference rasterizer was only able to draw axis-aligned quads
which prevented the gl_VertexID tests from working. Implement a
triangle rasterizer instead.

This was tested by running the functional draw tests that compare the
GPU output with a software-rasterized one. Chrome stable has the same
errors with and without this patch.

This was also tested by running the simple_reference_context tests.
There was one failure but pen and paper rasterization showed that the
test did not account for some precision issues so it was modified. On
the same test drawQuad and drawTriangles produce very different results
but pen and paper rasterization shows that drawTriangles has the
correct output.