Improve global variable initialization test
Add a test where a ternary operation condition references a uniform,
and another test where a uniform struct is assigned to a global.

An option to specify uniform values for GLSL tests is added for the
second test. This could be also used to refactor other existing tests
that currently are not using the GLSL test helpers because they need
to set uniforms.