Add a skip list to the dEQP test runner.
This allows tests to be skipped if they are failing because of
insurmountable graphics driver bugs. Deliberately, entries apply to all
operating systems, browsers and GPU types. The mechanism is deliberately
orthogonal to the existing support for the "filter" query variable.

The implementation is unfortunately complicated because of confusing
mutual recursion between some of the inner methods of tcuTestCase, and
the desire to report reasonable warnings when tests are skipped.
Untangling these methods is left for future work; patches welcome.

Some initial failure suppressions are added in this pull request.
Happily, NVIDIA's recently-released 364.15 driver fixes at least some,
if not all, of these failures. The tests that are now passing will be
removed in the next pull request. (It was desired to have an example in
the version history of how to use the skip list.)

Regenerated deqp-deps.js via "build.py depfile". Did not run the Closure
typechecker on these changes; believe the script for running it needs to
be updated. (This is on my to-do list.)