Expand tests for S3TC compressed textures to cover TexStorage2D.
This is the first time that uninitialized compressed textures can be
exposed to the JavaScript level, and they must be tested. The new
tests assert that the texture contains either transparent or opaque
black, depending on the specific DXT internal format.

See http://crbug.com/605129 for fixes for this test in Chromium.