Add out-of-bounds-index-buffers.html conformance test
Out-of-bounds-index-buffers.html tests whether implementations successfully handle
the "Out-of-Range Array Accesses" section of the WebGL 1 specification.

The test accomplishes this by having a vertex shader with two attributes: a position
and value. The position attribute will come from large position vertex buffer
filled with tri-strips that form quads which cover the entire canvas. This vertex
buffer will not be indexed off the end. The value attribute will come from a tiny
vertex buffer that is filled with a known-ahead-of-time "random" value. This buffer
will be indexed off the end. The shader tests whether the random value is either
found in the vertex buffer or is the zero index. Upon success, the shader outputs
green. On failure, the shader outputs red.

The test passes if implementations either:
1) Flag INVALID_OPERATION and do not output anything to the canvas OR
2) Flag NO_ERROR and output green for all canvas pixels.

Future improvements can be made that test more data types than just FLOAT as well
datatypes with different numbers of components.