Fix a bug in rrRenderer.drawTriangles
In former implementation of drawTriangles, all points on the sides of a triangle are regarded as being inside the triangle, so points on the border of two triangles will be rendered twice, which leads to the problem that when blending is enabled, points on the common side of two triangles will be blended twice and get wrong color.
e.g. SharedColorbuffer tests in FboRenderTest

Based on native dEQP codes, this patch re-implements drawTriangle which specifies the inclusiveness of the sides of a triangle.