Add a test for tricky loop conditions
Some types of transformations in shader translation may need to add
new statements to implement certain types of expressions. In case an
expression inside a loop condition or loop expression requires this
kind of a transformation, it needs special handling. Add a test for
bugs in this area.

There are known failures in this test on ANGLE's DirectX backend and
NVIDIA's OpenGL drivers.