Interpolation qualifier in all linked shaders must match
GLSL ES 3.00.4 specifies that "The type and presence of the
interpolation qualifiers and storage qualifiers and invariant qualifiers
of variables with the same name declared in all linked shaders must
match, otherwise the link command will fail." in Section 4.3.9 Interpolation.

The restriction for interpolation qualifier is relaxed to "It is a
link-time error if, within the same stage, the interpolation qualifiers
of variables of the same name do not match." in Section 4.5
Interpolation Qualifiers, GLSLangSpec 4.40.