Fix conformance/context/context-lost.html to generate INVALID_OPERATION for lost WebGLObjects.
The spec says that when a method on the context or an extension is called
that doesn't [WebGLHandlesContextLoss], first generate INVALID_OPERATION
if any object args have their 'invalidated flag' set. If either that, or
IsContextLost() is true, return defaults, etc.

The spec language does not appear to address what to do about methods
called on extensions which have been disabled/detached. This is not
covered in this test, though.