Improve tricky loop conditions test
Using exact float comparisons makes some of the tests fail on D3D once
the other bugs are fixed, so add a little bit of tolerance to the
float comparisons.

WebGL has also made it illegal to use the sequence operator on arrays,
so remove the incorrect test for that.

One test description is also clarified.