Add new test to expose issues in texture-offset-out-of-range
In texture-offset-out-of-range the shader uses a uniform variable to
represent the texture coordination, which is not a common usage and will
lead to chromium WebGL2 context crash on Windows with Intel driver on
Broadwell machines.

This patch uses attributes instead of uniform variables to work around
this issue in texture-offset-out-of-range and add new tests to expose all
related issues explicitly.