Split gl-uniform-arrays.html to avoid timeout
There are a lot of calls of gl.getUniform() in gl-uniform-arrays.html,
which are very time consuming on Chromium in debug version and will
make the test as a single file spend quite a long time to finish.

See http://crbug.com/603906 for more info.

This patch intends to split this test into small files to fix this issue.