Make compressed texture max size test reliable
ArrayBuffer allocations larger than 1 GB are not reliable on 32-bit
browsers. To make the compressed texture size conformance test
reliable, the max allocation size it uses needs to be limited down to
1 GB from the theoretical limit it was previously using.

To restore as much of the lost coverage as possible, cover max texture
size with 2D textures that are large only in one dimension and small
in the other dimension.

A separate stress test that uses ArrayBuffers up to the theoretical 4
GB limit is also added. This is valuable to have since it can reveal
bugs in 64-bit browsers that the lighter conformance test does not
reveal anymore. It also uses larger cube map textures in its positive
tests, raising the size limit from 2K x 2K to 8K x 8K.