Remove preferLowPowerToHighPerformance
This is being removed in favor of "performancePreference". Since only
Safari has shipped with support for this flag, but doesn't honor the
requested value, it is safe to remove. No content should be depending on it.

An other option would be to define how this value is converted to the
new performancePreference.