Added test from Andre Weissflog (@floooh) for draw buffers hang on macOS. (#2362)
This was originally reported under http://crbug.com/696187 and was
already fixed by the time it was reported to Apple via Radar.

Confirmed that this test runs properly on drivers that don't contain
the bug, and doesn't execute properly (causes the browser to lock up)
on drivers that do. There will be no false positives.