Add test for srgb emulation of generateMipmap (#2382)
When we call generateMipmaps for srgb texture, we need to emulate under the
hood for desktop GL prior to GL 4.4. The tests in this pull request are used
to expose errors during the emulation in Chromium:

1.During the emulation of generateMipmap for srgb texture, we tried to
generate all mipmaps of the original texture. But if the texture is immutable
, we can't generate mipmaps exceed the original range. See the bug at
https://crbug.com/712096.
2.In addition, during the emulation, there are some errors if the width and
height are not equal.