Add the concept of an opaque multiview framebuffer to WEBGL_multiview
An opaque multiview framebuffer acts as a multiview framebuffer, but
has immutable attachments defined by another web API. Under the hood,
the views of an opaque multiview framebuffer could be implemented in
whichever way is most optimal for a given situation. For example, they
could be texture layers, or side-by-side viewports on a 2D texture.

This will make it possible to create a zero-copy display pipeline that
makes use of multiview rendering. Another web API such as WebVR 2.0
could expose an opaque multiview framebuffer for rendering, and
specify rules when it gets displayed or invalidated. Since the
framebuffer is owned entirely by the other web API and only interacts
with WebGL in the role of a render target, the interactions between
WebGL and the other API remain fairly simple.

The opaque multiview framebuffer would be possible to support in
WebGL 1.0 as well, since it does not require API user facing support
for texture arrays. This revision doesn't yet fully address WebGL 1.0
compatibility of the extension, but is intended as one step towards
that.

This spec also leaves open the possibility to expose the underlying
memory of an opaque multiview framebuffer in some other way as well.
It could be complemented by a separate API for rendering to only one
of the views at a time, for example.

Includes a fix to an example in the WEBGL_multiview spec.