Fix dEQP multisample test tolerance for sample counts >= 64 (#2412)
Previously the test wasn't able to handle sample counts >= 64. This
was because it's counting unique colors, and it was detecting colors
with <= 3 difference in one channel as the same. But if you have 65
different gray levels between 0 and 255, at least two of them will be
set apart by only 3.

This made the test fail on Quadro drivers that supported 64 samples.

Now the test automatically lowers the tolerance if it is required by
having a high sample count.

After this fix the test at least has a chance of handling high sample
counts correctly, though it still needs to be verified if this fix is
sufficient on its own.