Revise spec on untouched fragment outputs to match implementations
At least on ANGLE-based WebGL implementations, fragment color is
always written into the framebuffer if gl_FragColor or gl_FragData is
statically used in the shader.

Since this doesn't seem to be causing any problems, the easiest way
to fix this is to change the spec to match the implementation.