Roll back shader restrictions in WEBGL_multiview spec (#2471)
From prototyping we have learned that most of the expected perfomance
gains from the extension can be realized without the shader
restrictions on NVIDIA OpenGL drivers, and particularly in the context
of the web the shader restrictions might only buy marginal additional
performance. As such it makes sense to leave the restrictions out to
make the API simpler and easier to program.