Require a multiview program to draw to a multiview framebuffer
Supporting drawing to multiview framebuffers with non-multiview
programs could require generating multiple versions of the program
object behind the scenes. This would require more complex code and
would result in worse and less predictable performance. Since this
kind of use has limited utility, it is better to make it to generate
an error.