Clarify WEBGL_multiview spec
The multiview spec is now written against the latest revision of the
native OVR_multiview spec. The native OVR_multiview spec has been
tightened to specify more error conditions, so some of the addendums
in the WebGL spec are not necessary anymore.

Formatting and examples are also improved.

This doesn't introduce any major functional changes.