Add tests for blitting to cover width/height overflow.
The range of an integer is [-0x7fffffff - 1, 0x7fffffff].
We restrict the width/height to the max value that can be
store in an integer in latest WebGL 2 spec. If we
reverse the data order for blitting, an invalid value like
0x7fffffff + 1 would become a valid value for an integer:
-0x7fffffff - 1. But it is invalid for blitting.

The new added tests cover this corner case to make it clear.