Specify binding one buffer to two transform feedback outputs
This case used to be underspecified, so behavior varies between
platforms. Prohibit having the same buffer bound to two transform
feedback outputs when calling BeginTransformFeeback to avoid
non-portable behaviors from now on.

Currently the test doesn't yet generate the expected error on any of
the platforms tested. On Chromium or Firefox running on ANGLE's D3D
backend, the buffer is not touched by transform feedback.
On Chromium's GL backend, only one of the outputs is written to the
buffer by transform feedback.

The behavior is the same on recent HW/drivers from both NVIDIA and
Intel.