Fix WebGL 2.0 spec about non-zero divisor requirement
The conformance tests were previously fixed to remove the requirement
that there's at least one non-zero divisor, as it is not needed in
D3D11-backed implementations of WebGL. This change brings the spec in
line with the tests.

The test is also expanded a bit to cover drawArrays and drawElements,
not just their instanced variants.

Currently Chromium passes the test, but Firefox follows the old 2.0.0
spec rather than the newer version of the test.