Explicitly set image's src in tex-video-using-tex-unit-non-zero.html .
It's not clear whether there's any guarantee that an img tag whose
source is a data: URL will be fully loaded before inline scripts below
it execute. Instead, set an onload handler for the image and set its
src in script.

Intended to address intermittent failures in this test per
http://crbug.com/830901 .