Add references to OffscreenCanvas.
OffscreenCanvas is documented in the WHATWG's current specifications,
and browsers are beginning to ship support for fetching WebGL contexts
against them, as well as use them as texture sources. Document this
properly in the WebGL specification.

In Chromium, this work was tracked under issue http://crbug.com/563816
and many dependent bugs.