Add a proposal for a BPTC compressed texture extension (#2710)
The proposed extension simply exposes the native compressed texture
extension ARB_texture_compression_bptc.

The extension should be supportable on both OpenGL and D3D11. In D3D11
the compressed texture formats exposed here are called BC7 and BC6H.
D3D11 also has signed and unsigned variants of the float format and
linear and SRGB variants of the fixed point format.

The extension is possible to support on both WebGL 1.0 and 2.0.