Improve compressed texture tests
A new test utility collection for compressed texture tests is added in
sdk/tests/js/tests/compressed-texture-utils.js. It can validate common
functionality in compressed texture extensions, such as enumerating
formats and size restrictions on data that is passed in to texture
uploads.

This also removes duplicate testing of querying
COMPRESSED_TEXTURE_FORMATS when no texture compression extensions are
enabled. This is tested in textures/misc/compressed-tex-image.html so
it doesn't need to be tested in individual compressed texture
extension tests. compressed-tex-image.html is now enabled also for
WebGL 2.0.

New coverage is added for trying to use all the different compressed
texture formats when the extensions are disabled.

Coverage is added also for trying to use more varied incorrect
dimensions for ASTC textures.