Bikeshedding on WEBGL_video_texture
This is a placeholder PR for bikeshedding on WEBGL_video_texture. I've suggested a rename here, because I think the old one was confusing.

I like something like this because it retains the "texImage2D" heritage. Without it, it's not clear that this behaves much like an upload call. However I'm open to other ideas.