Clarify texStorage*D initialization for compressed formats
The reference is made to ES 3.0 section 3.8.4 pseudocode, which in
turn is prefaced with

  If internalformat is a compressed texture format, then references to
  TexImage* should be replaced by CompressedTexImage*, with format, type
  and data replaced by any valid imageSize and data.

Include a clarification to this effect directly in the WebGL spec to
avoid requiring the reader to dig through to see if online compression
of a "zero" color is implied.

This is a copy of commit d0612c40c67 for the latest file.