Amend ANGLE_multiview to allow for paused XFB.
Rationale: a paused XFB object is essentially invisible to the
rendering engine. The base ES spec says: "When transform feedback
is paused, no vertices are recorded." Moreover, many operations
that fail when an XFB is active are allowed when it is also paused.
This change makes the multiview behaviour consistent with the rest
of the spec.