glBindBufferRange with size param less than 4 generates INVALID_VALUE
According to section 2.12.6.3 Uniform Buffer Object Storage of
the OpenGL ES 3.0 spec, the smallest element of a uniform block,
a bool, is still read as a typed uint (4 bytes).

Binding a buffer of size less than 4 since it is not large enough
to even fill the smallest possible uniform block would generate
INVALID_OPERATION at draw time. It should be reasonable to add
validation that the size param is not less than 4.

It is impossible to have a uniform block size that is not
a multiple of 4. Regardless, the spec allows calling
glBindBufferRange with a size that is not a multiple of 4 (You
can have an 11-byte buffer range for a 8-byte uniform block).
Some drivers incorrectly validate that the size param when the
target is gl.UNIFORM_BUFFER must be a multiple of 4. To workaround
this, size can be rounded down to the nearest multiple of 4 so the
driver does not generate an error.

Rounding down a value less than 4 to the nearest multiple of 4
would produce 0 which is also an error if passed to
glBindBufferRange. Specing that size >= 4 makes it safe to always
round size down.