Remove Program Pipeline from WebGL 2.0 compute
Program Pipeline is not a core feature for compute shader, it is
not very important. Furthermore, program pipeline is not very
useful. We didn't find that some demos or applications rely on it
and state that it is a must-to-have feature. In addition, we didn't
see a simple way to implement this feature on Direct3D 11. So we tend
to remove this feature from WebGL 2.0 compute.